ns: move IECommerceInterface
